Revert "Fix most gcc and clang warnings (#7)"
This reverts commit a0aeb369aac897b89b1a76e9f09d07d6d319f987.

+++ Containers/sg_OwnedMap.hpp
@@ -209,7 +209,7 @@ std::vector<KeyType> OwnedMap<KeyType, ValueType, Capacity>::getKeys() const noe
 {
     JUCE_ASSERT_MESSAGE_THREAD;
     std::vector<KeyType> result{};
-    result.reserve(static_cast<size_t>(size()));
+    result.reserve(size());
     for (auto const & node : *this) {
         result.push_back(node.key);
     }
